refactor: refactoring wallet-migration modals
