+++ packages/frontend/src/components/common/TwoFactorDisableBanner.js
@@ -8,7 +8,7 @@ import { selectAccountSlice } from '../../redux/slices/account';
 import WalletClass, { wallet } from '../../utils/wallet';
 import AlertTriangleIcon from '../svg/AlertTriangleIcon';
 import LockIcon from '../svg/LockIcon';
-import Disable2FAModal from '../wallet-migration/Disable2FA';
+import Disable2FAModal from '../wallet-migration/modals/Disable2faModal/Disable2FA';
 import FormButton from './FormButton';
 
 
@@ -124,7 +124,7 @@ export default function TwoFactorDisableBanner() {
             const accountsKeyTypes = await Promise.all(
                 accounts.map(getAccountWithAccessKeysAndType)
             );
-    
+
             setAccounts(accountsKeyTypes.reduce(((acc, { accountId, keyType }) => keyType === WalletClass.KEY_TYPES.MULTISIG ? [...acc, accountId] : acc), []));
         };
         if (loadedAccounts.length > 0 && accounts.sort() !== loadedAccounts.sort()) {
@@ -144,7 +144,7 @@ export default function TwoFactorDisableBanner() {
             </div>
             <div className='content'>
                 <h4 className='title'>
-                    { accountsCount }
+                    {accountsCount}
                     {' '}
                     {
                         accountsCount > 1
@@ -160,7 +160,7 @@ export default function TwoFactorDisableBanner() {
                 onClick={showModal}
                 color='red'
             >
-                <LockIcon color='#FEF2F2' /> 
+                <LockIcon color='#FEF2F2' />
                 <Translate id='twoFactorDisbleBanner.button' />
             </FormButton>
             {
