+++ packages/frontend/src/components/wallet-migration/WalletMigration.jsx
@@ -1,58 +1,24 @@
 import React, { useCallback, useEffect } from 'react';
-import { useSelector } from 'react-redux';
 
-import { ACCOUNT_ID_SUFFIX } from '../../config';
-import { selectAvailableAccounts } from '../../redux/slices/availableAccounts';
-import { encodeAccountsToHash, generatePublicKey, keyToString } from '../../utils/encoding';
-import { getLedgerHDPath } from '../../utils/localStorage';
-import { wallet } from '../../utils/wallet';
-import Disable2FAModal from './Disable2FA';
-import MigrateAccounts from './MigrateAccounts';
-import MigrationSecret from './MigrationSecret';
-import RotateKeysModal from './modals/RotateKeysModal';
-import SelectDestinationWallet from './SelectDestinationWallet';
+import Disable2FAModal from './modals/Disable2faModal/Disable2FA';
+import ExportAccountsModal from './modals/ExportKeysModal/ExportAccountsModal';
+import RotateKeysModal from './modals/RotateKeysModal/RotateKeysModal';
+
 
 export const WALLET_MIGRATION_VIEWS = {
-    MIGRATION_SECRET: 'MIGRATION_SECRET',
-    SELECT_DESTINATION_WALLET: 'SELECT_DESTINATION_WALLET',
-    MIGRATE_ACCOUNTS: 'MIGRATE_ACCOUNTS',
     DISABLE_2FA: 'DISABLE_2FA',
-    ROTATE_KEYS: 'ROTATE_KEYS'
+    ROTATE_KEYS: 'ROTATE_KEYS',
+    MIGRATE_ACCOUNTS: 'MIGRATE_ACCOUNTS'
 };
 
 const initialState = {
     activeView: null,
     wallet: null,
-    migrationKey: generatePublicKey()
 };
 
-const getAccountsData = async (accounts) => {
-    const accountsData = [];
-    for (let i = 0; i < accounts.length; i++) {
-        const accountId = accounts[i];
-        const keyPair = await wallet.getLocalKeyPair(accountId);
-        accountsData.push([
-            accountId,
-            keyPair?.secretKey || '',
-            getLedgerHDPath(accountId),
-        ]);
-    }
-
-    return accountsData;
-};
-
-const encodeAccountsToURL = async (accounts, publicKey, { getUrl }) => {
-    const accountsData = await getAccountsData(accounts);
-    const hash = encodeAccountsToHash(accountsData, publicKey);
-    const networkId = ACCOUNT_ID_SUFFIX === 'near' ? 'mainnet' : 'testnet';
-    const href = getUrl({ hash, networkId });
-
-    return href;
-};
 
 const WalletMigration = ({ open, onClose }) => {
     const [state, setState] = React.useState(initialState);
-    const availableAccounts = useSelector(selectAvailableAccounts);
 
     const handleStateUpdate = (newState) => {
         setState({...state, ...newState});
@@ -66,27 +32,10 @@ const WalletMigration = ({ open, onClose }) => {
         handleStateUpdate({ activeView });
     }, [handleStateUpdate]);
 
-    const showMigrationPrompt = useCallback(() => {
-        handleSetActiveView(WALLET_MIGRATION_VIEWS.SELECT_DESTINATION_WALLET);
-    }, [handleSetActiveView]);
-
-    const showMigrateAccount = useCallback(() => {
-        handleSetActiveView(WALLET_MIGRATION_VIEWS.MIGRATE_ACCOUNTS);
-    }, [handleSetActiveView]);
-
-    const onContinue = useCallback(async () => {
-        let url = '';
-        url = await encodeAccountsToURL(
-            availableAccounts,
-            state.migrationKey,
-            state.wallet
-        );
-        window.open(url, '_blank');
-    }, [state.migrationKey, availableAccounts, state.wallet]);
-
+  
     useEffect(() => {
         if (open) {
-            handleSetActiveView(WALLET_MIGRATION_VIEWS.ROTATE_KEYS);
+            handleSetActiveView(WALLET_MIGRATION_VIEWS.MIGRATION_SECRET);
         } else {
             handleSetActiveView(null);
         }
@@ -98,38 +47,26 @@ const WalletMigration = ({ open, onClose }) => {
                 <Disable2FAModal
                     onClose={onClose}
                     handleSetActiveView={handleSetActiveView}
+                    data-test-id="disable2FAModal"
                 />
             )}
             {state.activeView === WALLET_MIGRATION_VIEWS.ROTATE_KEYS && (
                 <RotateKeysModal
                     onClose={onClose}
                     handleSetActiveView={handleSetActiveView}
+                    data-test-id="rotateKeysModal"
                 />
             )}
-            {
-                state.activeView === WALLET_MIGRATION_VIEWS.MIGRATION_SECRET && (
-                    <MigrationSecret
-                        showMigrationPrompt={showMigrationPrompt}
-                        showMigrateAccount={showMigrateAccount}
-                        secretKey={keyToString(initialState.migrationKey)}
-                    />
-                )}
-            {state.activeView === WALLET_MIGRATION_VIEWS.SELECT_DESTINATION_WALLET && (
-                <SelectDestinationWallet
-                    wallet={state.wallet}
+            {state.activeView === WALLET_MIGRATION_VIEWS.MIGRATION_SECRET && (
+                <ExportAccountsModal
                     onClose={onClose}
-                    handleSetWallet={handleSetWallet}
                     handleSetActiveView={handleSetActiveView}
+                    handleSetWallet={handleSetWallet}
+                    state={state}
+                    data-test-id="exportAccountsModal"
                 />
             )}
-            {
-                state.activeView === WALLET_MIGRATION_VIEWS.MIGRATE_ACCOUNTS && (
-                    <MigrateAccounts
-                        accounts={availableAccounts}
-                        onContinue={onContinue}
-                        onClose={onClose}
-                    />
-                )}
+            
         </div>
     );
 };
